refactor(editor): extract selection property commands
Move reference/value projection visibility, component identity writes, MOS target structure edits, and annotation actions out of App into the Properties feature. Move instance label lookup ownership to Instance Display while retaining the previous helper export for compatibility.

Validation: pnpm typecheck; pnpm ci:static; 26 focused App/helper/instance-display/netlist-authoring unit tests; focused Playwright contracts for single/group Reference visibility and reviewed SKY130 Model switching; pnpm gate:plan for owned paths; git diff --check.

Test-Impact: no-test-change — behavior is unchanged and covered by existing unit contracts plus focused Properties browser tests for both document and structure transaction paths.
